change data type

+++ src/java/business/SavingAccount.java
@@ -27,7 +27,7 @@ public class SavingAccount extends Account implements Serializable {
     @Column(nullable = false)
     private int minBalance;
     @Column(nullable = false)
-    private int savingAmount;
+    private Double savingAmount;
 
     public String getSavingAccountId() {
         return savingAccountId;
@@ -45,11 +45,11 @@ public class SavingAccount extends Account implements Serializable {
         this.minBalance = minBalance;
     }
 
-    public int getSavingAmount() {
+    public Double getSavingAmount() {
         return savingAmount;
     }
 
-    public void setSavingAmount(int savingAmount) {
+    public void setSavingAmount(Double savingAmount) {
         this.savingAmount = savingAmount;
     }
